Begin restructuring CalendarConversation class
Lots of comments have been added, some variables renamed, and a few
bits of code are broken out into their own functions.

Our position in the list can be stored as an int in the session, but
intents and code for changing position in the list have not been
added yet, and we still return a tell response and end the session,
rather than an ask response which keeps the session active.

The first intent request we get in a session, the position session
attribute is initialized to 0.  Later in the session, the position
may be changed when the user expresses an intent to move forward or
backward in the list.