Use enums instead of a bunch of random strings
Enums let us take advantage of Java's static typing to provide
compile-time guarantees about the correctness of our code.

As an example of this, we can now add new intents to the enum and be
sure that it will also be added to the supportedIntentNames list, as the
constructor is able to iterate over the set of all enumerated values to
add them to the list.

We also are able to take advantage of the fact that Java enums are
classes that can have methods, in order to implement common patterns as
methods of the enum.

For example, we can now get the session state attribute from a session
by calling `SessionState.valueOf(session)` and doing one null check,
rather than repeating the complicated series of gets and checks required
previously.