Initial commit of chainable filters
An easy to use query filtering system is implemented using Scala and a
"Functional Relational Mapper" (FRM), Slick.  Slick has a code generator
that was able to generate a functional model of the existing database.
Slick takes care of the hard part of generating SQL for arbitrary query
filters.

A simple set of classes representing different query filters (category,
start time, event title) can be collected in a list and persisted in the
session for a conversation, and applied in order to a query.  This
allows us to create functionality for a user to specify selection
criteria in any ad-hoc fashion.

The implementation at this point is still a work in progress.