Switch to FilterChain in session
Using a FilterChain rather than a List<Filter> in the conversation
eliminates some awkward conversions and simplifies the mapping to and
from JSON objects for use in the session attributes.

JSON mapping is implemented in FilterChain with a constructor that takes
the amorphous blob we get out of the session, and a .toAttrib() method
that emits a mappable object that can be saved in the session.

The session attribute for the filters is changed to FILTER_CHAIN.

FilterChain now has an .append(Filter) method that returns a new
FilterChain with the passed Filter appended to the end of the list.
This avoids mutation while still allowing a FilterChain to be extended.